Move GC to standard library
Currently, Alloy's GC lives in `library/alloc` and is then re-exported
to `library/std` which is the convention for most heap allocated
types. Unfortunately, this means that it is unable to use standard
library items such as Mutexes and Channels as code in `library/alloc`
must be `#[no_std]` compliant.

By moving `Gc` into `library/std` we can use standard library types
for our off-thread finalisation implementation. This commit simply moves
code around and does not make any functional changes to Alloy.
